+++ mix/test.cldr.ex
@@ -3,6 +3,6 @@ defmodule Money.Cldr do
 
   use Cldr,
     locales: ["en", "de", "it", "es"],
-    providers: [Cldr, Cldr.Numbers]
+    providers: [Cldr, Cldr.Number]
 
 end